move png_animated into png file
